requirements.txt
autopep8==1.5.1
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
html5lib==1.0.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.3
jsons==1.1.2
MarkupSafe==1.1.1
pycodestyle==2.5.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
typish==1.5.3
urllib3== 1.25.11
webencodings==0.5.1
Werkzeug==1.0.1
